Fixed GAL22V10 programming

+++ database.h
@@ -40,7 +40,7 @@ typedef struct gal_config {
   uint8_t row_width;     // how many bytes a row have
   uint16_t ues_address;  // user electronic signature address
   uint8_t ues_size;      // ues size in bits
-  uint8_t acw_address;   // address of 'architecture control word'
+  uint16_t acw_address;   // address of 'architecture control word'
   uint8_t acw_size;      // acw size in bits
   uint16_t *acw_bits;    // acw bits order
 } gal_config_t;
