+++ tl866iiplus.c
@@ -476,7 +476,7 @@ int tl866iiplus_unlock_tsop48(minipro_handle_t *handle, uint8_t *status) {
   return EXIT_SUCCESS;
 }
 int tl866iiplus_write_jedec_row(minipro_handle_t *handle, uint8_t *buffer,
-                                uint8_t row, size_t size) {
+                                uint16_t row, size_t size) {
   uint8_t msg[64];
   memset(msg, 0, sizeof(msg));
   msg[0] = TL866IIPLUS_WRITE_JEDEC;
@@ -488,7 +488,7 @@ int tl866iiplus_write_jedec_row(minipro_handle_t *handle, uint8_t *buffer,
 }
 
 int tl866iiplus_read_jedec_row(minipro_handle_t *handle, uint8_t *buffer,
-                               uint8_t row, size_t size) {
+                               uint16_t row, size_t size) {
   uint8_t msg[32];
   memset(msg, 0, sizeof(msg));
   msg[0] = TL866IIPLUS_READ_JEDEC;
